Fix field name for query parameters
Currently the fields for server requests corresponding to query
parameters are generated using the same algorithm than the names of the
query parameters as used in the HTTP protocol. For example, if the name
of the parameter is `size` then the name of the field is also `size`.
But this will not work for parameter names with more than one word. For
example, if a new `MyParameter` is added then it will result in a field
name that is `my_parameter` instead of the expected `myParameter`. This
patch fixes that.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>