Don't generate empty set of constants for errors
Currently the code generator generates an empty `const (...)` when there
are no errors defined in the API. This will be removed by `gofmt` but
it is frequent to forget to run it when the SDK is updated, and as a
result those pull requests are rejected. To avoid that this patch
changes the code so that it doesn't generate that empty `const`.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>