Use private model
Currently the project uses the model from the `ocm-api-model` project
for tests. This complicates adding tests that require new features in
the model, as those features have to be added to that project first. It
also means that updating that project may break this one. To avoid that
this patch adds a minimal model inside this repository.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>