Improve map support
This patch improves the support for attributes whose type is a map where
the key is a string and the value is an struct. This means that now
types like the following can be used:

```
struct AccessToken {
	Auths [string]Auth
}

struct Auth {
	Username string
	Email string
}
```

The JSON representation for these types will be like this:

```json
{
  "auths": {
    "my.com": {
      "username": "myuser",
      "email": "mymail"
    },
    "your.com": {
      "username": "youruser",
      "email": "yourmail"
    }
}
```

In the generated code these types will be built like this:

```go
token, err := amv1.NewAccessToken().
	Auths(map[string]*amv1.AuthBuilder{
		"my.com": amv1.NewAuth().
			Username("myuser").
			Email("mymail"),
		"your.com": amv1.NewAuth().
			Username("youruser").
			Email("yourmail"),
	}).
	Build()
if err != nil {
	...
}
```

Signed-off-by: Juan Hernandez <jhernand@redhat.com>