Stateless adapters
Currently the adapters that convert HTTP requests into requests to the
implementations of th servers are stateful: they contain a reference to
the server that they use. This means that each request requires the
allocation of a new adapter for each path segment. To avoid that
unnecessary allocations this patch changes the adapters so that they are
stateless functions instead of objects.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>