Send not found when server returns nil target
Currently the HTTP adaters don't check if the value returned by locators
is `nil`. That results in a panic when that happens. To avoid that this
patch changes the generated code so that will check if the returned
value is `nil` and will send a 404 error response if it is.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>