Don't fail on wrong kind
Currently reading an object fails if the `kind` attribute isn't one of
the expected ones. But the server tends to report this incorrectly, and
the SDK doesn't really need the `kind` attribute to figure out the type
of the object. To make the SDK more tolerant of server inconsistences
this patch changes the generated code so that it will not require a
correct value for the `kind` attribute.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>