Fix generation of OpenAPI paths
Currently the code that generates OpenAPI paths uses the `append`
function to extend the generated paths inside an iterative algorithm.
This results in some of the generated paths (which are slices of
pointers to objects) sharing the same underlying array. When the slice
is modified via any of those slides that share the underlying array the
generated path overwrites an existing one, resulting in duplicated and
missing paths. This patch addresses that issue replacing that use of
`append` with a combination of `make` and `copy`.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>