Add search method
The metamodel does not support requesting a list of items through a POST
method. And the List method only supports GET requests. To allow for
a search request using a POST, we add a new method Search. It takes the
same parameters as a List (page, size) as well as a Body. It returns
a list of Items as well as the collection parameters (page, size,
total).