Add presence bitmap
This patch changes the generated code so that instead of using pointers to
store attributes of scalar values (strings, integers, etc) it stores them
directly and uses a bitmap to keep track of which attributes are present. This
reduces the use of CPU and the number of allocations. In a test that loads 1
million clusters in memory it reduces CPU by approx 20% and memory by approx
10%.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>