Add annotations
This patch adds initial support for annotation model concepts. In the
language annotations will be written like this:

```
// An annotation to change the name used for this attribute in the
// Go generated code:
@go(name = "MyAttr")
attribute MyAttribute Integer

// An annotation to mark an attribute as deprecated:
@deprecated
attribute MyAttribute Integer
```

Note that both these `@go` and `@deprecated` annotations are just
examples of the syntax and what could be their use. This patch just adds
the basic support to the language, but no special meaning for those two
annotations.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>