Add metrics support generator
This patch adds a new code generator that generates the
`path_tree_data.go` file containing the tree of URL paths that the
metrics round tripper of the SDK will use to calculate the `path` label
of metrics. This means that it is no longer necessary to use the
`X-Metric` header to carry the value of that label, so the patch also
removes all uses of that header.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>