Check loops in locator paths
Currently the model parser doesn't check for loops in the graph of
locators. This can cause infinite loops or stack overflows in tools that
process it, in particular in the tool that generates the tree of URL
segments used for metrics. To avoid that this patch introduces an
additional check for that. When it detects a loop it print an error
message and abort the processing. For example, with a simple recursive
locator like this:

```
resource A {
        locator A {
                target A
        }
}
```

It will print the following error message:

```
E:Locator 'A' introduces a loop, it is defined in resource 'MyService.V1.A' and targets that same resource
```

For a more compliated loop like this:

```
resource A {
        locator B {
                target B
        }
}

resource B {
        locator C {
                target C
        }
}

resource C {
        locator A {
                target A
        }
}
```

It will generate an error message with the summary and additional
information messages explaining the details of the loop:

```
E: Locator 'A' defined in resource 'MyService.V1.C' introduces a loop, see the details below
I: First locator of the loop is 'B' defined in resource 'MyService.V1.A' and targeting resource 'MyService.V1.B'
I: Next locator of the loop is 'C' defined in resource 'MyService.V1.B' and targeting resource 'MyService.V1.C'
I: Last locator of the loop is 'A' defined in resource 'MyService.V1.C' and targeting resource 'MyService.V1.A'
```

Related: https://github.com/openshift-online/ocm-api-metamodel/issues/135
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>