Check for `io.EOF` before trying to parse response body
Currently the clients always try to parse the HTTP response body, even
when it is empty. The `Unmarshal...` methods used for that compare the
reader to `http.NoBody` and return nil in that case. This isn't reliable
because transport wrappers may use a different type to indicate an empty
response body. To avoid that this patch replaces that check with a
buffered reader. This is used to peek the first byte of the response
body and then check if it results in `io.EOF`. This approach should work
regardless of the type of the response body.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>