Add `@json` and `@http` annotations
The `@json` annotation is used to set an alternative name for attributes
and parameters in JSON representations. For example, if an attribute is
defined line this:

```
attribute CPUID string
```

The generated code will use `cpuid` in the JSON representation. This may
be incorrect if the intent was to use two separate words `CPU` and `ID`.
the annotation can be used to tell the code generator to use `cpu_id`
instead:

```
@json(name = "cpu_id")
attribute CPUID string
```

The `@http` annotation is similar, but intended for segments of URLs
corresponding to methods, locators and query parameters.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>