Don't parse names
Currently we parse the names and separate them into words, so that it is
easier to reconstruct that later using camel case or snake case. That
adds a lot of complication and means that the names used in the `.model`
files aren't the names actually used in the JSON representation. In
order to be able to change that in the future this patch changes the
parser so that it uses the names from the `.model` files without parsing
or changing them.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>