Don't transform names
Currently the tool transforms names of attributes and parametes from
camel case to snake case in order to use them as the names of the JSON
fields, HTTP query parameters and URL segments. This is confusing for
developers creating the model. To avoid that this patch changes the tool
so that it uses those names without changes.

Note that this means that existing models will have to be updated to
replace the camel case names with snake case names.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>