Make annotation helpers public
Currently the functions used to get the `name` parameter of the `@json`,
`@http` and `@go` annotations are private in the packages that use it.
This makes it difficult to use them in other packages, in particular
they can't be used in the OpenAPI generator package, and that is
necessary in order to fix a bug. This patch moves those functions to a
new `annotations` package and makes them public. The OpenAPI generator
bug will be fixed in a separate commit.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>